move 'in_tracing_context' to util crate and use in 'app' crate
