+++ bin/spiced/src/main.rs
@@ -19,6 +19,7 @@ use opentelemetry::global;
 use rustls::crypto::{self, CryptoProvider};
 use telemetry::noop::NoopMeterProvider;
 use tokio::runtime::Runtime;
+use util::in_tracing_context;
 
 #[cfg(feature = "alloc-jemalloc")]
 #[global_allocator]
@@ -81,7 +82,7 @@ fn main() {
     }
 
     if let Err(err) = tokio_runtime.block_on(start_runtime(args)) {
-        runtime::in_tracing_context(|| {
+        in_tracing_context(|| {
             tracing::error!("{err}");
         });
     }
@@ -93,7 +94,7 @@ fn main() {
 }
 
 async fn start_runtime(args: spiced::Args) -> Result<(), Box<dyn std::error::Error>> {
-    runtime::in_tracing_context(|| {
+    in_tracing_context(|| {
         if let Some(allocator_name) = get_allocator_name() {
             tracing::info!(
                 "Starting runtime {version} (allocator: {allocator_name})",
