+++ crates/app/src/lib.rs
@@ -38,6 +38,7 @@ use spicepod::{
     },
     extension::Extension,
 };
+use util::in_tracing_context;
 
 pub mod runtime;
 
@@ -374,15 +375,19 @@ impl AppBuilder {
             }
 
             if dependent_spicepod.runtime != Runtime::default() {
-                tracing::warn!(
-                    "Spicepod dependency '{dependency}' has 'runtime' field(s) defined. Runtime configuration must be set in primary spicepod. '{dependency}' runtime configuration will be ignored."
-                );
+                in_tracing_context(|| {
+                    tracing::warn!(
+                        "Spicepod dependency '{dependency}' has 'runtime' field(s) defined. Runtime configuration must be set in primary spicepod. '{dependency}' runtime configuration will be ignored."
+                    );
+                });
             }
 
             if dependent_spicepod.management.is_some() {
-                tracing::warn!(
-                    "Spicepod dependency '{dependency}' has 'management' field(s) defined. Management configuration must be set in primary spicepod. '{dependency}' management configuration will be ignored."
-                );
+                in_tracing_context(|| {
+                    tracing::warn!(
+                        "Spicepod dependency '{dependency}' has 'management' field(s) defined. Management configuration must be set in primary spicepod. '{dependency}' management configuration will be ignored."
+                    );
+                });
             }
 
             spicepods.push(dependent_spicepod);
