+++ crates/runtime/src/builder.rs
@@ -36,6 +36,7 @@ use app::App;
 use spicepod::component::caching::Caching;
 use token_provider::registry::TokenProviderRegistry;
 use tokio::sync::{Mutex, RwLock};
+use util::in_tracing_context;
 
 type DatafusionConfigurationCallback = fn(&mut DataFusion);
 
@@ -176,7 +177,7 @@ impl RuntimeBuilder {
             .as_ref()
             .and_then(|app| app.runtime.results_cache.clone())
         {
-            crate::in_tracing_context(|| {
+            in_tracing_context(|| {
                 tracing::warn!(
                     "The `results_cache` Runtime parameter is deprecated and will be removed in a future release. Use `caching.sql_results` instead. For more information, visit: https://spiceai.org/docs/features/caching"
                 );
@@ -312,7 +313,7 @@ fn parse_memory_limit(memory_limit: Option<String>) -> Option<u64> {
         .map(|v| v.get_adjusted_unit(byte_unit::Unit::B).get_value() as u64);
 
     if memory_limit.is_none() {
-        crate::in_tracing_context(|| {
+        in_tracing_context(|| {
             tracing::warn!(
                 "An invalid Runtime memory limit was specified: {original_memory_limit} A memory limit must be specified as an integer in GB, MB, or KB size."
             );
@@ -320,7 +321,7 @@ fn parse_memory_limit(memory_limit: Option<String>) -> Option<u64> {
     }
 
     if memory_limit == Some(0) {
-        crate::in_tracing_context(|| {
+        in_tracing_context(|| {
             tracing::warn!(
                 "A Runtime memory limit of 0 was specified: {original_memory_limit} A memory limit must be greater than 0."
             );
