+++ crates/util/src/lib.rs
@@ -23,7 +23,9 @@ use std::{
 pub mod fibonacci_backoff;
 pub use backoff::Error as RetryError;
 pub use backoff::future::retry;
+mod tracing_util;
 use tokio::{sync::oneshot, time::Instant};
+pub use tracing_util::in_tracing_context;
 
 #[allow(clippy::cast_precision_loss)]
 #[allow(clippy::cast_sign_loss)]
